requirements.txt
flask
flask-cors
textblob
nltk
requests
beautifulsoup4
python-docx
PyPDF2
reportlab
vaderSentiment